iec61850_mms: rename to iec61850_client
