+++ lib/nodes/CMakeLists.txt
@@ -83,7 +83,7 @@ endif()
 
 # Enable IEC61850 node-types when libiec61850 is available
 if(WITH_NODE_IEC61850)
-    list(APPEND NODE_SRC iec61850_sv.cpp iec61850_mms.cpp iec61850.cpp)
+    list(APPEND NODE_SRC iec61850_sv.cpp iec61850_client.cpp iec61850.cpp)
     list(APPEND LIBRARIES PkgConfig::LIBIEC61850)
 endif()
